moved bloglist button to component file for functionality.
