+++ src/App.js
@@ -11,8 +11,8 @@ import AppLayout from "./appLayout/AppLayout";
 import Profile from './pages/Profile';
 import AboutPage from "./pages/AboutPage";
 import Login from "./components/Login";
-import BlogListButton from './pages/BlogListButton'; // Make sure this path matches your file structure
-
+import BlogListButton from './components/BlogListButton'; // Make sure this path matches your file structure
+import BlogList from './pages/BlogList';
 import HomePage from './pages/Home';
 import BlogPage from "./pages/BlogPage";
 import StoryList from './components/StoryList'; // Import the new StoryList component
@@ -21,7 +21,7 @@ import StoryList from './components/StoryList'; // Import the new StoryList comp
 const router = createBrowserRouter(createRoutesFromElements(
   <Route path="/" element={<AppLayout />}>
     <Route path="/blogCreation" element={<Profile />} />
-    <Route path="/blogList" element={<BlogListButton />} />   
+    <Route path="/blogList" element={<BlogList />} />   
     <Route path="/about" element={<AboutPage />} />
     <Route path="/homePage" element={<HomePage />} />
     <Route path="/blogPage" element={<BlogPage />} />
