+++ src/pages/BlogList.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-
+import { Pencil, Trash2, X, Check, Plus } from 'lucide-react';
 function BlogList() {
     const [blogs, setBlogs] = useState([
         { title: 'My First Blog', content: 'This is my first post.' },
